Address Selection: Saving Selected User Address For Checkout
whenever user selects the address, we are updating it into the Redux store

and we are storing it as selected user checkout address.
Why?

Because on checkout on the next step we will be needing this address.

And on the payment page we will be needing the selected checkout address of the user.

So that is what we are storing.

And we are also using this to update the UI as you are seeing over here.